Determine Whether an Indefinite Article is Needed for a Particular Class

This function returns the correct indefinite article based on the class of the input.

Usage

dis_indefinite(x, indefinite_class = NULL)

Arguments

x

Required object; a parameter argument to test.

indefinite_class

Optional character vector; if the class of an object is listed here, and class(x) matches this class, the function will return "an".

Value

If the input is an integer or the class of the input is included in indefinite_class, the function will return "an". Otherwise, the function will return "a".

Examples

dis_indefinite(x = "test")
#> [1] "a"
dis_indefinite(x = 2)
#> [1] "a"